backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
flasgger==0.9.7.1
python-jose==3.5.0
requests==2.32.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-Migrate==4.0.1
alembic==1.11.1
psycopg[binary]==3.2.9
Pillow==10.4.0
supabase==2.7.4
PyJWT[crypto]==2.8.0